lndservices: move invoices_client.go into the new package
